pyproject.toml
[project]
name = "MediSage-AI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "google-genai>=1.70.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.14",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
